items: port Item_Initialise to TRX
This in turn resolves #2904 as we use the same check as was done in TR2
regarding enemy hitpoints.
