+++ docs/tr1/README.md
@@ -486,6 +486,7 @@ Not all options are turned on by default. Refer to `TR1X_ConfigTool.exe` for det
 - fixed the game crashing if a cinematic is triggered but the level contains no cinematic frames
 - fixed pushblocks being rotated when Lara grabs them, most noticeable if asymmetric textures have been used
 - fixed Lara becoming clamped if she picks up an item under a steeply sloped ceiling
+- fixed enemy hitpoints being doubled in demo mode as a result of NG+
 
 #### Cheats
 - added a fly cheat
